Consolidate Layout components
